datagen: Fix linter for `Read` from `"math/rand"`
The function `math/rand.Read` was marked as deprecated, but we are keeping it
because it's faster, and we don't care about the cryptographic qualities of
this data.
`math/rand.Seek` was also marked as deprecated, and we don't need it,
as the package now automatically seeds the global random number generator.

Signed-off-by: Tatiana Nesterenko <tatiana@nspcc.io>